[PATCH v3] hwmon: Add occ_i2c HWMON driver - Revise code based on comments from Jeremy Kerr <jk@ozlabs.org>. - Add new HWMON attributes "update_interval" and "name". - Change driver name form occ.c to occ_i2c.c
